+++ dev-elements/test/__mocks__/pixi.js.js
@@ -8,6 +8,7 @@
 
 export class DisplayObject {
   setParent = (f: Container) => f;
+  toLocal = (point: Point) => point;
 }
 export class Point {
   x: number;
@@ -24,17 +25,21 @@ export class Point {
 export class Container extends DisplayObject {
   height: number;
   pivot: Point;
+  rotation: number;
+  scale: Point;
   width: number;
 
   constructor() {
     super();
 
     this.pivot = new Point();
+    this.rotation = 0;
+    this.scale = new Point;
     this.height = 200;
     this.width = 200;
   }
   addChild = (f: Container) => f;
-  getChildAt = (f: Container) => f;
+  getChildAt = () => this;
 }
 
 function returnThis() {