+++ src/engine/components/numberGenerator/index.js
@@ -4,16 +4,21 @@
  *
  * @flow
  */
-import { Text, TextStyle } from 'pixi.js';
-import { TweenLite, TimelineLite, Linear } from 'gsap';
+import { type Container, Point, Text, TextStyle } from 'pixi.js';
+import { TweenLite, TimelineLite, Linear, Power1, SlowMo } from 'gsap';
 
 import { HALF, ONE, ZERO, TWO, ANCHOR_CENTER, ACTOR_MOVEMENT_DURATION } from 'constants/numbers';
 import { UnableToLeaveTheNumericLine } from 'engine/helpers/errors';
 import { EASE_BE_HAPPY, EASE_BE_SAD } from 'engine/helpers/customEases';
 import { getRandomFloat } from 'helpers/randomizers';
 
+const THREE = 3;
 const SIX = 6;
 const SHAKE_DISTANCE = 2.3;
+const FALL_DISTANCE_MULTIPLIER = 0.95;
+const FALL_SPIN_DURATION = 3;
+const JUMP_DURATION = 0.4;
+const SQUISH_DURATION = 2.5;
 
 export const START_STATE = 'start';
 export const STOP_STATE = 'stop';
@@ -35,8 +40,9 @@ export type NumberActor = {|
   finalPosition: string,
   beHappy(state: ActorEmotionState): void,
   beSad(state: ActorEmotionState): void,
+  beTheFallenOne(): Promise<void>,
   changeActor(number: number): void,
-  hasEnteredToNumericLine(): Promise<void>,
+  hasEnteredToNumericLine(emptyBlock: Container): Promise<void>,
   resetPosition(): void,
   updatePosition(newPosition: string): Promise<void>,
 |};
@@ -47,6 +53,34 @@ export type StaticNumberActor = {|
 |};
 export type ActorEmotionState = 'start' | 'stop';
 
+const beTheFallenOneConfig = (
+  view: Text,
+  size: number,
+) => (function beTheFallenOne(): Promise<void> {
+  /**
+   * Has fell into oblivion
+   */
+  this.position = undefined;
+
+  return new Promise((onComplete) => {
+    const timeline = new TimelineLite({ onComplete });
+
+    timeline
+      .to(view, ACTOR_MOVEMENT_DURATION, {
+        y: view.y - size * FALL_DISTANCE_MULTIPLIER,
+        ease: Linear.easeNone,
+      })
+      .to(view, FALL_SPIN_DURATION, {
+        rotation: 50,
+        ease: Power1.easeIn,
+      }, ZERO)
+      .to(view.scale, SQUISH_DURATION, {
+        x: 0,
+        y: 0,
+        ease: Power1.easeIn,
+      }, ACTOR_MOVEMENT_DURATION);
+  });
+});
 const emotionConfig = (
   view: Text,
   size: number,
@@ -86,12 +120,17 @@ const emotionConfig = (
 export const staticNumberGenerator = (number: number, size: number): StaticNumberActor => {
   const style = new TextStyle({
     ...styleRaw,
-    fontSize: size / HALF + size / SIX,
+    fontSize: size + size / THREE,
   });
   const view = new Text(number.toString(), style);
 
   view.anchor.set(ANCHOR_CENTER);
   view.x = view.y = size / HALF;
+  /**
+   * PixiJS recomendation for better text resolution
+   * @see https://github.com/pixijs/pixi.js/wiki/Performance-Tips#text
+   */
+  view.scale.x = view.scale.y = ANCHOR_CENTER;
 
   return {
     view,
@@ -100,27 +139,36 @@ export const staticNumberGenerator = (number: number, size: number): StaticNumbe
   };
 };
 
+const GLOBAL_POINT = new Point();
 const hasEnteredToNumericLineConfig = (
   view: Text,
   size: number,
-  margin: number,
   /**
    * This function needs the number's scope. That's why is a named function.
    *
-   * @return {Promise<void>} animation promise
+   * @param {Container} emptyBlock  who is going to be the new parent
+   * @return {Promise<void>}        animation promise
    */
-) => (function hasEnteredToNumericLine(): Promise<void> {
+) => (function hasEnteredToNumericLine(emptyBlock: Container): Promise<void> {
   this.position = undefined;
+  const localPosition = emptyBlock.toLocal(GLOBAL_POINT, view);
 
-  view.x = size / HALF;
-  view.y = size + view.x + margin + margin;
+  view.setParent(emptyBlock);
+  view.position = localPosition;
 
   return new Promise((onComplete) => {
-    TweenLite.to(view, ACTOR_MOVEMENT_DURATION, {
-      y: size / HALF,
-      ease: Linear.easeNone,
-      onComplete,
-    });
+    const timeline = new TimelineLite({ onComplete });
+
+    timeline
+      .to(view, ACTOR_MOVEMENT_DURATION, {
+        y: size / HALF,
+        ease: Linear.easeNone,
+      })
+      .to(view.scale, ACTOR_MOVEMENT_DURATION, {
+        x: view.scale.x * TWO,
+        y: view.scale.y * TWO,
+        ease: SlowMo.ease.config(JUMP_DURATION, ZERO, true),
+      }, ZERO);
   });
 });
 
@@ -160,6 +208,12 @@ const resetPositionConfig = (
 
   view.x = initialPosition[0] * (size + margin) + size / HALF + margin;
   view.y = initialPosition[1] * (size + margin) + size / HALF + margin;
+  /**
+   * PixiJS recomendation for better text resolution
+   * @see https://github.com/pixijs/pixi.js/wiki/Performance-Tips#text
+   */
+  view.scale.x = view.scale.y = ANCHOR_CENTER;
+  view.rotation = 0;
 });
 const changeActorConfig = (
   view: Text,
@@ -188,7 +242,7 @@ const numberGenerator = (
 ): NumberActor => {
   const style = new TextStyle({
     ...styleRaw,
-    fontSize: size / HALF + size / SIX,
+    fontSize: size + size / THREE,
   });
   const view = new Text(number.toString(), style);
   const initialPosition = position.split(',').map((string: string): number => (parseInt(string)));
@@ -197,6 +251,11 @@ const numberGenerator = (
 
   view.x = initialPosition[0] * (size + margin) + size / HALF + margin;
   view.y = initialPosition[1] * (size + margin) + size / HALF + margin;
+  /**
+   * PixiJS recomendation for better text resolution
+   * @see https://github.com/pixijs/pixi.js/wiki/Performance-Tips#text
+   */
+  view.scale.x = view.scale.y = ANCHOR_CENTER;
 
   return {
     view,
@@ -204,8 +263,9 @@ const numberGenerator = (
     finalPosition,
     beHappy: emotionConfig(view, size, true),
     beSad: emotionConfig(view, size, false),
+    beTheFallenOne: beTheFallenOneConfig(view, size),
     changeActor: changeActorConfig(view),
-    hasEnteredToNumericLine: hasEnteredToNumericLineConfig(view, size, margin),
+    hasEnteredToNumericLine: hasEnteredToNumericLineConfig(view, size),
     resetPosition: resetPositionConfig(view, initialPosition, size, margin),
     updatePosition: updatePositionConfig(view, size, margin),
   };